Small fix in gmail agent and conversation
Fix gmail unnecessary extra fields when there's no html detected.

POINT OF IMPROVEMENT: Make the code that checks if the email is an html. Make it accurate that it can ignore small signs of what html detector it detects.

Then fixed conversation_agent that now reads and returns to the user the final context after the plan execution.

Made improvements as well in filtering the final context if multi-step were to be executed. The final context would be a huge file in itself. So fixed what fields are truly necessary for the user as user don't really care about the backend side of things for the summarization of the final context or what was done.